requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
pandas
pymysql
APScheduler
openpyxl
cryptography
